+++ cmd/template/cluster/provider/templates/azure/template.go
@@ -2,7 +2,6 @@ package azure
 
 import (
 	_ "embed"
-	"strings"
 )
 
 //go:embed cluster.yaml.tmpl
@@ -18,13 +17,13 @@ var kubeadmControlPlane string
 var azureMachineTemplate string
 
 // GetTemplate merges all .tmpl files.
-func GetTemplate() string {
-	return strings.Join([]string{
+func GetTemplates() []string {
+	// Order is important here.
+	// The order in this slice determines in which order files will be applied.
+	return []string{
 		cluster,
 		azureCluster,
 		kubeadmControlPlane,
 		azureMachineTemplate,
-		// Adds a separator at the end of the joined template for easier joinability of cluster and node pool templates.
-		"",
-	}, "\n---\n")
+	}
 }